update ImageCaption evaluate
